refactor: :truck: separate landing components
