+++ app/[locale]/page.tsx
@@ -1,20 +1,11 @@
-import Brand from '@/components/Brand'
-import VideoOverlay from '@/components/VideoOverlay'
-
-const src = {
-  mp4: '/assets/video/v.mp4',
-  mov: '/assets/video/v.mov',
-  ogv: '/assets/video/v.ogv',
-}
+import ContentGrid from '@/components/ContentGrid'
+import Landing from '@/components/Landing'
 
 export default function Home() {
   return (
-    <main className="relative flex h-screen items-center justify-center bg-black">
-      <Brand
-        src="/assets/logo.png"
-        alt="toyz logo"
-      />
-      <VideoOverlay src={src} />
+    <main>
+      <Landing />
+      <ContentGrid />
     </main>
   )
 }
